% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postion.R
\name{addPositionHorizontal}
\alias{addPositionHorizontal}
\title{addPositionHorizontal}
\usage{
addPositionHorizontal(
  geometryName,
  geometry,
  n,
  factor = 0.1,
  left_offset = 1000,
  range = c(1, 5),
  rowBreak = TRUE
)
}
\arguments{
\item{rowBreak}{}
}
\description{
addPositionHorizontal
}
