% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postion.R
\name{geom_sf_arrange}
\alias{geom_sf_arrange}
\title{Convenient geom for arranged geometries}
\usage{
geom_sf_arrange(
  spacing = 0.1,
  offset = 1000,
  scale_range = c(1, 5),
  wrap_rows = TRUE,
  ...
)
}
\arguments{
\item{spacing}{Spacing factor}

\item{offset}{Initial offset}

\item{scale_range}{Scaling range}

\item{wrap_rows}{Whether to wrap rows}

\item{...}{Additional arguments passed to geom_sf}
}
\description{
Convenient geom for arranged geometries
}
