% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postion.R
\name{geom_sf_sample}
\alias{geom_sf_sample}
\title{Convenient geom for sampled points}
\usage{
geom_sf_sample(n = 10, ...)
}
\arguments{
\item{n}{Number of points to sample}

\item{...}{Additional arguments passed to geom_sf}
}
\description{
Convenient geom for sampled points
}
