% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrwGemLine}
\alias{nrwGemLine}
\title{North Rhine-Westphalia Municipality Line Data}
\format{
An sf object with 396 rows and 5 columns:
\describe{
\item{geometry}{Line geometry for municipality boundaries}
}
}
\source{
PALAVA-App (LWL und LVR)
}
\usage{
nrwGemLine
}
\description{
Spatial line data for municipalities in North Rhine-Westphalia, Germany.
Contains geometric boundaries as line features for all 396 municipalities.
}
\examples{
\dontrun{
library(ggplot2)
library(sf)

ggplot(nrwGemLine) +
  geom_sf() +
  theme_palava_map() +
  labs(title = "NRW Municipality Boundaries") +
  labs_palava()
}
}
\keyword{datasets}
