% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrwGemPolygons}
\alias{nrwGemPolygons}
\title{North Rhine-Westphalia Municipality Polygon Data}
\format{
An sf object with 396 rows and 5 columns:
\describe{
\item{geometry}{Polygon geometry for municipality areas}
}
}
\source{
PALAVA-App (LWL und LVR)
}
\usage{
nrwGemPolygons
}
\description{
Spatial polygon data for municipalities in North Rhine-Westphalia, Germany.
Contains area boundaries for all 396 municipalities.
}
\examples{
\dontrun{
library(ggplot2)
library(sf)

ggplot(nrwGemPolygons) +
  geom_sf(aes(fill = factor(1:n()))) +
  theme_palava_map() +
  scale_fill_palava_discrete() +
  labs(title = "NRW Municipality Polygons") +
  labs_palava() +
  theme(legend.position = "none")
}
}
\keyword{datasets}
