% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postion.R
\name{position_sample}
\alias{position_sample}
\title{Position: Sample Points Within Geometries}
\usage{
position_sample(n = 10)
}
\arguments{
\item{n}{Number of points to sample per geometry (default: 10)}
}
\description{
Samples random points within each geometry, creating a jittered effect
}
