% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_palava_continuous}
\alias{scale_color_palava_continuous}
\title{scale_color_palava_continuous}
\usage{
scale_color_palava_continuous(..., aesthetics = "color")
}
\arguments{
\item{aesthetics}{}
}
\description{
scale_color_palava_continuous
}
