% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_palava_discrete}
\alias{scale_color_palava_discrete}
\title{scale_color_palava_discrete}
\usage{
scale_color_palava_discrete(..., aesthetics = "color")
}
\arguments{
\item{aesthetics}{}
}
\description{
scale_color_palava_discrete
}
