% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_palava_discrete}
\alias{scale_fill_palava_discrete}
\title{scale_fill_palava_discrete}
\usage{
scale_fill_palava_discrete(..., aesthetics = "fill")
}
\arguments{
\item{aesthetics}{}
}
\description{
scale_fill_palava_discrete
}
