% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{palava_colors_discrete}
\alias{palava_colors_discrete}
\title{palava_colors_discrete
set up colors for discrete color scale}
\usage{
palava_colors_discrete(n)
}
\arguments{
\item{n}{}
}
\description{
palava_colors_discrete
set up colors for discrete color scale
}
