% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_palava_continuous}
\alias{scale_fill_palava_continuous}
\title{scale_fill_palava_continuous}
\usage{
scale_fill_palava_continuous(..., aesthetics = "fill")
}
\arguments{
\item{aesthetics}{}
}
\description{
scale_fill_palava_continuous
}
