% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_palava_continuous}
\alias{scale_fill_palava_continuous}
\title{scale_fill_palava_continuous}
\usage{
scale_fill_palava_continuous(..., aesthetics = "fill")
}
\arguments{
\item{...}{Arguments passed to ggplot2::continuous_scale()}

\item{aesthetics}{Character string or vector of aesthetics to apply the scale to}
}
\value{
A ggplot2 scale object for continuous fill aesthetics
}
\description{
scale_fill_palava_continuous
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(x = waiting, y = eruptions, fill = density)) +
  geom_raster() +
  scale_fill_palava_continuous()
}
